requirements.txt
streamlit>=1.30.0
httpx>=0.25.0
pyyaml>=6.0
jsonschema>=4.20.0
deepdiff>=6.7.0
prance>=23.6.21.0
openapi-spec-validator>=0.7.1
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
